(PC-40648) refactor(home): extract video module header in a specific component
